requirements.txt
asgiref==3.5.0
Django==4.0.2
django-filter==21.1
django-multiselectfield==0.1.12
gunicorn==20.1.0
sqlparse==0.4.2
tzdata==2021.5
whitenoise==6.0.0
